requirements.txt
# AutoGen Council v2.4.0 Requirements
# Production-ready multi-agent AI system dependencies

# Core ML/AI Libraries - Week 1 Lock-in
torch==2.3.0+cu121
transformers==4.41.0
sentencepiece>=0.1.97
sentence-transformers==3.2.0
sympy>=1.11.0
numpy>=1.24.0

# API & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.1

# Performance Boosters - Week 1 Lock-in
faiss-cpu>=1.7.3          # For vector search (use faiss-gpu if you have CUDA)
accelerate>=0.20.0        # For faster model loading
flash-attn==2.3.6         # Flash Attention for latency optimization
xformers==0.0.37          # Memory-efficient transformers

# Logic Programming (Optional)
# pyswip>=0.2.10          # Uncomment if SWI-Prolog is installed

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
python-dateutil>=2.8.0
pathlib2>=2.3.0

# HTTP client for LLM endpoints
aiohttp>=3.8.0
httpx==0.25.2
requests==2.31.0

# Data processing
pandas==2.1.4

# Async support
asyncio-mqtt>=0.11.0

# Logging and monitoring
prometheus-client>=0.16.0
structlog==23.2.0

# Configuration
python-dotenv==1.0.0
pyyaml>=6.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional: Math and ML libraries (for local skills)
sympy==1.12
scipy==1.11.4

# Phase 3 Memory Integration
redis>=4.5.0
redis[hiredis]==5.0.4     # High-performance Redis client for pattern mining
qdrant-client>=1.6.0

# Pattern Mining & Clustering (Spiral v2.7.0)
sentence-transformers==2.6.1   # Lightweight embedding model for clustering  
hdbscan==0.8.33                # Density-based clustering for prompt patterns
rdbtools==0.1.20               # Redis RDB file parser for traffic snapshots

# Phase 4 TensorRT-LLM Integration  
aiohttp>=3.8.0            # Async HTTP client for TensorRT backend
locust>=2.17.0            # Load testing framework

# Docker and containerization support
docker>=6.1.0             # Docker SDK for Python
docker-compose>=1.29.2    # Docker Compose support 